updated buying/selling logic
currently have it so that BUY(and sell) doesnt even do a call to quote (not even cache), because i dont see any point in it. its not so much dual case, as it is just get the price when they commit. i know in class he said there has to be a quote first, but i dont see any upside, since were willing to just change the price on them anyways